# docs/swagger.yaml
basePath: /api
definitions:
...
  handler.PageData:
    properties:
      list: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  handler.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  handler.serviceAdjustBalanceReq:
    properties:
      delta:
        type: integer
      remark:
        type: string
    required:
    - delta
    type: object
  handler.serviceConsumeReq:
    properties:
      amount:
        type: integer
      key:
        type: string
    required:
    - key
    type: object
host: localhost:8080
info:
  contact: {}
  description: 卡密管理系统 API 文档
  title: CloudKey API
  version: "1.0"
paths:
...
  /service/consume-result:
    get:
      description: 通过 X-Service-Key 认证，根据 request_id 查询消费或调额的结果
      parameters:
      - description: 请求ID
        in: query
        name: request_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 操作结果
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 缺少 request_id 参数
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 服务账号密钥无效
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: 未找到对应的操作记录
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ServiceKeyAuth: []
      summary: 根据请求ID查询操作结果
      tags:
      - 服务账号API
  /service/keys:
    get:
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: '状态过滤: active/exhausted/disabled/expired'
        in: query
        name: status
        type: string
      - description: 别名前缀搜索
        in: query
        name: alias
        type: string
      - description: 后缀精准搜索
        in: query
        name: key_suffix
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 分页卡密列表
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.PageData'
              type: object
        "401":
          description: 服务账号密钥无效
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ServiceKeyAuth: []
      summary: 服务账号查询卡密列表
      tags:
      - 服务账号API
    post:
      consumes:
      - application/json
      description: 通过 X-Service-Key 认证，服务账号创建卡密
      parameters:
      - description: 卡密参数
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 服务账号密钥无效
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ServiceKeyAuth: []
      summary: 服务账号创建卡密
      tags:
      - 服务账号API
  /service/keys/{id}:
    delete:
      parameters:
      - description: 卡密ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 无效的卡密 ID
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 服务账号密钥无效
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ServiceKeyAuth: []
      summary: 服务账号删除卡密
      tags:
      - 服务账号API
    get:
      parameters:
      - description: 卡密ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 卡密详情
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 无效的卡密 ID
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 服务账号密钥无效
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: 卡密不存在
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ServiceKeyAuth: []
      summary: 服务账号查询卡密详情
      tags:
      - 服务账号API
    patch:
      consumes:
      - application/json
      parameters:
      - description: 卡密ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新字段
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 服务账号密钥无效
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ServiceKeyAuth: []
      summary: 服务账号更新卡密
      tags:
      - 服务账号API
  /service/keys/{id}/adjust-balance:
    post:
      consumes:
      - application/json
      description: 通过 X-Service-Key 认证，增加或减少卡密额度，所有变动记录在流转日志中
      parameters:
      - description: 卡密ID
        in: path
        name: id
        required: true
        type: integer
      - description: 调整参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.serviceAdjustBalanceReq'
      produces:
      - application/json
      responses:
        "200":
          description: 调整结果
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 服务账号密钥无效
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ServiceKeyAuth: []
      summary: 服务账号调整卡密额度
      tags:
      - 服务账号API
  /service/keys/{id}/disable:
    patch:
      parameters:
      - description: 卡密ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 禁用成功
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 无效的卡密 ID
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 服务账号密钥无效
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ServiceKeyAuth: []
      summary: 服务账号禁用卡密
      tags:
      - 服务账号API
  /service/keys/{id}/enable:
    patch:
      parameters:
      - description: 卡密ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 启用成功
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 无效的卡密 ID
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 服务账号密钥无效
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ServiceKeyAuth: []
      summary: 服务账号启用卡密
      tags:
      - 服务账号API
  /service/keys/consume:
    post:
      consumes:
      - application/json
      description: 通过 X-Service-Key 认证，扣减指定卡密的剩余额度
      parameters:
      - description: 扣减参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.serviceConsumeReq'
      produces:
      - application/json
      responses:
        "200":
          description: 扣减结果
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 参数错误或卡密无效
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 服务账号密钥无效
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ServiceKeyAuth: []
      summary: 服务账号扣减卡密额度
      tags:
      - 服务账号API
  /service/keys/export:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 导出数据
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 服务账号密钥无效
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ServiceKeyAuth: []
      summary: 服务账号导出卡密（文本格式）
      tags:
      - 服务账号API
  /service/keys/export/json:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 导出数据 JSON 数组
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 服务账号密钥无效
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ServiceKeyAuth: []
      summary: 服务账号导出卡密（JSON 格式）
      tags:
      - 服务账号API
  /service/keys/status:
    get:
      description: 通过 X-Service-Key 认证，根据卡密值查询状态
      parameters:
      - description: 卡密值
        in: query
        name: sk
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 卡密状态信息
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 缺少卡密参数
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 服务账号密钥无效
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: 卡密不存在
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ServiceKeyAuth: []
      summary: 服务账号查询卡密状态
      tags:
      - 服务账号API
...
  /tenant/service-accounts:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 服务账号列表
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: 服务账号列表
      tags:
      - 租户-服务账号
    post:
      consumes:
      - application/json
      parameters:
      - description: 服务账号参数
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  properties:
                    created_at:
                      type: string
                    id:
                      type: integer
                    is_active:
                      type: boolean
                    name:
                      type: string
                    raw_key:
                      type: string
                  type: object
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: 创建服务账号
      tags:
      - 租户-服务账号
  /tenant/service-accounts/{id}:
    delete:
      parameters:
      - description: 服务账号ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 无效的服务账号 ID
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: 删除服务账号
      tags:
      - 租户-服务账号
  /tenant/service-accounts/{id}/toggle:
    patch:
      consumes:
      - application/json
      parameters:
      - description: 服务账号ID
        in: path
        name: id
        required: true
        type: integer
      - description: 状态
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 操作成功
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: 启用/禁用服务账号
      tags:
      - 租户-服务账号
...
securityDefinitions:
  ApiKeyAuth:
    description: 'Bearer JWT token, 格式: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
  ServiceKeyAuth:
    description: 服务账号密钥
    in: header
    name: X-Service-Key
    type: apiKey
swagger: "2.0"